third_party/blink/web_tests/fast/alignment/parse-align-content-expected.txt

This is a testharness.js-based test.
[FAIL] Test getting align-content values previously set through CSS.
  assert_equals: alignContent is not what is should. expected "last baseline" but got "normal"
[FAIL] Test getting and setting align-content through JS
  assert_equals: alignContent specified value is not what it should. expected "last baseline" but got "baseline"
Harness: the test ran to completion.
